requirements.txt
llama-cpp-python==0.2.11
sentence-transformers==2.2.2
fastapi==0.104.1
uvicorn==0.23.2
python-multipart==0.0.6
pydantic==2.4.2
langchain==0.0.335
langchain-community==0.0.16
pypdf==3.17.0
docx2txt==0.8
tiktoken==0.5.1
pandas==2.0.3
openpyxl==3.1.2
gunicorn==21.2.0
